created node directory
